refactor: LottoNumber caching전략 추가, LottoNumberGenerator변경
