+++ src/main/java/camp/nextcamp/edu/lotto/entity/LottoNumber.java
@@ -1,21 +1,50 @@
 package camp.nextcamp.edu.lotto.entity;
 
+import static camp.nextcamp.edu.lotto.constants.LottoConstants.*;
 import static camp.nextcamp.edu.lotto.exception.UserExceptionMesssage.*;
 import static camp.nextcamp.edu.lotto.module.LottoValidator.*;
 
+import java.util.Map;
 import java.util.Objects;
+import java.util.Random;
+import java.util.Set;
+import java.util.TreeSet;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 import camp.nextcamp.edu.lotto.exception.UserException;
 
 public class LottoNumber implements Comparable<LottoNumber> {
 	private final int number;
 
-	public LottoNumber(int number) {
-		validateNumber(number);
+	private LottoNumber(int number) {
 		this.number = number;
 	}
 
-	public void validateNumber(int number) {
+	private static class LottoNumberCache {
+		private static final Map<Integer, LottoNumber> numbers;
+
+		static {
+			numbers = IntStream.rangeClosed(LOTTO_MIN, LOTTO_MAX)
+				.boxed()
+				.collect(Collectors.toMap(i -> i, LottoNumber::new));
+		}
+	}
+
+	public static Set<LottoNumber> generateRandomNumber(Random random) {
+		return random.ints(LOTTO_MIN, LOTTO_MAX)
+			.distinct()
+			.limit(LOTTO_COUNT)
+			.mapToObj(LottoNumber::valueOf)
+			.collect(Collectors.toCollection(TreeSet::new));
+	}
+
+	public static LottoNumber valueOf(int number) {
+		validateNumber(number);
+		return LottoNumberCache.numbers.get(number);
+	}
+
+	private static void validateNumber(int number) {
 		if (!isInRange(number)) {
 			throw new UserException(IS_NOT_LOTTO_RANGE);
 		}
@@ -26,7 +55,7 @@ public class LottoNumber implements Comparable<LottoNumber> {
 		if (this == o) return true;
 		if (o == null || getClass() != o.getClass()) return false;
 		LottoNumber that = (LottoNumber) o;
-		return number == that.number;
+		return this == that;
 	}
 
 	@Override
