+++ src/test/java/camp/nextcamp/edu/lotto/LottoRandomGeneratorTest.java
@@ -9,13 +9,12 @@ import java.util.Set;
 import org.junit.jupiter.api.Test;
 
 import camp.nextcamp.edu.lotto.entity.LottoNumber;
-import camp.nextcamp.edu.lotto.module.LottoNumberGenerator;
 
 public class LottoRandomGeneratorTest {
 
 	@Test
 	void LottoRandomGenerator_확인() {
-		Set<LottoNumber> list = LottoNumberGenerator.generateRandomNumber(new Random());
+		Set<LottoNumber> list = LottoNumber.generateRandomNumber(new Random());
 
 		assertAll(
 			() -> assertThat(list.size())
